xmpp_transport: Reset encryption status when initiating new connection
If the transport has been previously connected, the isSecure flag may still
be set from a previous connection. This can lead to not negotiating TLS on the
new connection. In practice this caused a stream error, because the post-TLS
stream reset was still sent unconditionally.

This change makes reconnecting a transport negotiate TLS successfully.